feat: refactor browser-clients
