+++ examples/react-components/src/app/dashboard/page.tsx
@@ -46,7 +46,7 @@ import { Toaster, toast } from "sonner";
 
 export default function Dashboard() {
   const router = useRouter();
-  const { turnkey, authIframeClient, passkeyClient } = useTurnkey();
+  const { turnkey, passkeyIframeClient } = useTurnkey();
   const [loading, setLoading] = useState(true);
   const [accounts, setAccounts] = useState<any>([]);
   const [wallets, setWallets] = useState<any[]>([]);
@@ -87,7 +87,7 @@ export default function Dashboard() {
       suborgID: suborgId,
       otpType: OtpType.Email,
       contact: emailInput,
-      userIdentifier: authIframeClient?.iframePublicKey!,
+      userIdentifier: passkeyIframeClient?.iframePublicKey!,
     });
     if (!sendOtpResponse || !sendOtpResponse.otpId!) {
       toast.error("Failed to send OTP");
@@ -101,7 +101,7 @@ export default function Dashboard() {
       otpType: OtpType.Sms,
       contact: phoneInput,
       customSmsMessage: "Your Turnkey Demo OTP is {{.OtpCode}}",
-      userIdentifier: authIframeClient?.iframePublicKey!,
+      userIdentifier: passkeyIframeClient?.iframePublicKey!,
     });
     if (!sendOtpResponse || !sendOtpResponse.otpId!) {
       toast.error("Failed to send OTP");
@@ -132,8 +132,8 @@ export default function Dashboard() {
       toast.error("Email is already connected to another account");
       return;
     }
-
-    await authIframeClient?.addUserAuth({
+    await passkeyIframeClient?.updateUser({
+      organizationId: suborgId,
       userId: user.userId,
       email: emailInput,
     });
@@ -142,7 +142,7 @@ export default function Dashboard() {
       suborgID: suborgId,
       otpType: OtpType.Email,
       contact: emailInput,
-      userIdentifier: authIframeClient?.iframePublicKey!,
+      userIdentifier: passkeyIframeClient?.iframePublicKey!,
     });
     if (!sendOtpResponse || !sendOtpResponse.otpId!) {
       toast.error("Failed to send OTP");
@@ -166,8 +166,8 @@ export default function Dashboard() {
       toast.error("Phone Number is already connected to another account");
       return;
     }
-
-    await authIframeClient?.addUserAuth({
+    await passkeyIframeClient?.updateUser({
+      organizationId: suborgId,
       userId: user.userId,
       phoneNumber: phoneInput,
     });
@@ -177,7 +177,7 @@ export default function Dashboard() {
       otpType: OtpType.Sms,
       contact: phoneInput,
       customSmsMessage: "Your Turnkey Demo OTP is {{.OtpCode}}",
-      userIdentifier: authIframeClient?.iframePublicKey!,
+      userIdentifier: passkeyIframeClient?.iframePublicKey!,
     });
     if (!sendOtpResponse || !sendOtpResponse.otpId!) {
       toast.error("Failed to send OTP");
@@ -193,7 +193,7 @@ export default function Dashboard() {
     switch (oauthType) {
       case "Apple":
         oidcToken = await appleOidcToken({
-          iframePublicKey: authIframeClient?.iframePublicKey!,
+          iframePublicKey: passkeyIframeClient?.iframePublicKey!,
           clientId: process.env.NEXT_PUBLIC_APPLE_CLIENT_ID!,
           redirectURI: `${process.env
             .NEXT_PUBLIC_OAUTH_REDIRECT_URI!}dashboard`,
@@ -202,7 +202,7 @@ export default function Dashboard() {
 
       case "Facebook":
         oidcToken = await facebookOidcToken({
-          iframePublicKey: authIframeClient?.iframePublicKey!,
+          iframePublicKey: passkeyIframeClient?.iframePublicKey!,
           clientId: process.env.NEXT_PUBLIC_FACEBOOK_CLIENT_ID!,
           redirectURI: `${process.env
             .NEXT_PUBLIC_OAUTH_REDIRECT_URI!}dashboard`,
@@ -211,7 +211,7 @@ export default function Dashboard() {
 
       case "Google":
         oidcToken = await googleOidcToken({
-          iframePublicKey: authIframeClient?.iframePublicKey!,
+          iframePublicKey: passkeyIframeClient?.iframePublicKey!,
           clientId: process.env.NEXT_PUBLIC_GOOGLE_CLIENT_ID!,
           redirectURI: `${process.env
             .NEXT_PUBLIC_OAUTH_REDIRECT_URI!}dashboard`,
@@ -230,8 +230,8 @@ export default function Dashboard() {
         toast.error("Social login is already connected to another account");
         return;
       }
-
-      await authIframeClient?.addUserAuth({
+      await passkeyIframeClient?.createOauthProviders({
+        organizationId: suborgId,
         userId: user.userId,
         oauthProviders: [
           {
@@ -256,12 +256,13 @@ export default function Dashboard() {
       second: "2-digit",
     })}`;
     const { encodedChallenge, attestation } =
-      (await passkeyClient?.createUserPasskey({
+      (await passkeyIframeClient?.createUserPasskey({
         publicKey: { user: { name: siteInfo, displayName: siteInfo } },
       })) || {};
 
     if (encodedChallenge && attestation) {
-      await authIframeClient?.addUserAuth({
+      await passkeyIframeClient?.createAuthenticators({
+        organizationId: suborgId,
         userId: user.userId,
         authenticators: [
           {
@@ -285,7 +286,7 @@ export default function Dashboard() {
   };
 
   const handleDeleteAccount: any = async () => {
-    await authIframeClient?.deleteSubOrganization({
+    await passkeyIframeClient?.deleteSubOrganization({
       organizationId: suborgId,
       deleteWithoutExport: true,
     });
@@ -300,25 +301,25 @@ export default function Dashboard() {
   useEffect(() => {
     const manageSession = async () => {
       try {
-        if (turnkey && authIframeClient) {
+        if (turnkey && passkeyIframeClient) {
           const session = await turnkey?.getReadWriteSession();
           if (!session || Date.now() > session!.expiry) {
             await handleLogout();
           }
-          await authIframeClient.injectCredentialBundle(
+          await passkeyIframeClient.injectCredentialBundle(
             session!.credentialBundle
           );
-          const whoami = await authIframeClient?.getWhoami();
+          const whoami = await passkeyIframeClient?.getWhoami();
           const suborgId = whoami?.organizationId;
           setSuborgId(suborgId!);
 
-          const userResponse = await authIframeClient!.getUser({
+          const userResponse = await passkeyIframeClient!.getUser({
             organizationId: suborgId!,
             userId: whoami?.userId!,
           });
 
           setUser(userResponse.user);
-          const walletsResponse = await authIframeClient!.getWallets({
+          const walletsResponse = await passkeyIframeClient!.getWallets({
             organizationId: suborgId!,
           });
           setWallets(walletsResponse.wallets);
@@ -355,10 +356,11 @@ export default function Dashboard() {
             const defaultWalletId = walletsResponse.wallets[0].walletId;
             setSelectedWallet(defaultWalletId);
 
-            const accountsResponse = await authIframeClient!.getWalletAccounts({
-              organizationId: suborgId!,
-              walletId: defaultWalletId,
-            });
+            const accountsResponse =
+              await passkeyIframeClient!.getWalletAccounts({
+                organizationId: suborgId!,
+                walletId: defaultWalletId,
+              });
             setAccounts(accountsResponse.accounts);
             if (accountsResponse.accounts.length > 0) {
               setSelectedAccount(accountsResponse.accounts[0].address);
@@ -371,13 +373,13 @@ export default function Dashboard() {
         setLoading(false);
       }
     };
-    if (authIframeClient) {
+    if (passkeyIframeClient) {
       manageSession();
     }
-  }, [authIframeClient, turnkey]);
+  }, [passkeyIframeClient, turnkey]);
 
   const getWallets = async () => {
-    const walletsResponse = await authIframeClient!.getWallets({
+    const walletsResponse = await passkeyIframeClient!.getWallets({
       organizationId: suborgId!,
     });
     setWallets(walletsResponse.wallets);
@@ -405,7 +407,7 @@ export default function Dashboard() {
     setAnchorEl(null);
 
     // Fetch accounts for the selected wallet
-    const accountsResponse = await authIframeClient!.getWalletAccounts({
+    const accountsResponse = await passkeyIframeClient!.getWalletAccounts({
       organizationId: suborgId!,
       walletId,
     });
@@ -425,7 +427,7 @@ export default function Dashboard() {
           ? keccak256(toUtf8Bytes(messageToSign)) // Ethereum requires keccak256 hash
           : Buffer.from(messageToSign, "utf8").toString("hex"); // Solana doesn't require hashing
 
-      const resp = await authIframeClient?.signRawPayload({
+      const resp = await passkeyIframeClient?.signRawPayload({
         organizationId: suborgId!,
         signWith: selectedAccount!,
         payload: hashedMessage,
@@ -1091,7 +1093,7 @@ export default function Dashboard() {
               contact={emailInput ? emailInput : phoneInput}
               suborgId={suborgId}
               otpId={otpId!}
-              authIframeClient={authIframeClient!}
+              passkeyIframeClient={passkeyIframeClient!}
               onValidateSuccess={handleOtpSuccess}
               onResendCode={emailInput ? handleResendEmail : handleResendSms}
             />
