+++ packages/sdk-browser/src/__types__/base.ts
@@ -3,6 +3,10 @@ import type { WalletInterface, WalletStamper } from "@turnkey/wallet-stamper";
 import type * as SdkApiTypes from "../__generated__/sdk_api_types";
 import type { WebauthnStamper } from "@turnkey/webauthn-stamper";
 import type { IframeStamper } from "@turnkey/iframe-stamper";
+import type {
+  TurnkeyIframeClient,
+  TurnkeyPasskeyClient,
+} from "../__clients__/browser-clients";
 
 export type GrpcStatus = {
   message: string;
@@ -17,7 +21,7 @@ export enum MethodType {
 }
 
 export enum SessionType {
-  READ = "read",
+  READ_ONLY = "read_only",
   READ_WRITE = "read_write",
 }
 
@@ -93,6 +97,13 @@ export type TurnkeySDKClientConfig =
   | SDKClientConfigWithStamper
   | SDKClientConfigWithReadOnlySession;
 
+export interface TurnkeySDKClientPasskeyIframeConfig
+  extends BaseSDKClientConfig {
+  iframeStamper: IframeStamper;
+  passkeyStamper: WebauthnStamper;
+  readOnlySession: string;
+}
+
 export interface TurnkeySDKBrowserConfig {
   apiBaseUrl: string;
   defaultOrganizationId: string;
@@ -120,6 +131,26 @@ export interface IframeClientParams {
   dangerouslyOverrideIframeKeyTtl?: number;
 }
 
+export interface PasskeyClientParams {
+  rpId?: string;
+  timeout?: number;
+  userVerification?: UserVerificationRequirement;
+  allowCredentials?: PublicKeyCredentialDescriptor[];
+}
+
+export interface PasskeyIframeClientParams {
+  rpId?: string;
+  timeout?: number;
+  userVerification?: UserVerificationRequirement;
+  allowCredentials?: PublicKeyCredentialDescriptor[];
+  passkeyClient?: TurnkeyPasskeyClient;
+  iframeClient?: TurnkeyIframeClient;
+  iframeContainer?: HTMLElement | null | undefined;
+  iframeUrl?: string;
+  iframeElementId?: string;
+  readOnlySession?: string;
+}
+
 export interface TurnkeyWalletClientConfig extends SDKClientConfigWithStamper {
   wallet: WalletInterface;
 }
@@ -131,6 +162,7 @@ export enum AuthClient {
   Passkey = "passkey",
   Wallet = "wallet",
   Iframe = "iframe",
+  PasskeyIframe = "passkeyIframe",
 }
 
 export type TSessionResponse = Omit<
