+++ packages/sdk-browser/tsconfig.json
@@ -7,10 +7,6 @@
     "lib": ["dom"],
     "paths": {
       "@base-client": ["./src/__clients__/base-client"],
-      "@browser-client": ["./src/__clients__/browser-client"],
-      "@iframe-client": ["./src/__clients__/iframe-client"],
-      "@passkey-client": ["./src/__clients__/passkey-client"],
-      "@wallet-client": ["./src/__clients__/wallet-client"],
       "@constants": ["./src/constants"],
       "@models": ["./src/models"],
       "@storage": ["./src/storage"],
