+++ packages/sdk-react/src/components/auth/Auth.tsx
@@ -82,8 +82,7 @@ const Auth: React.FC<AuthProps> = ({
   customSmsMessage,
   customAccounts,
 }) => {
-  const { passkeyClient, authIframeClient, walletClient, turnkey } =
-    useTurnkey();
+  const { passkeyIframeClient, turnkey } = useTurnkey();
   const [email, setEmail] = useState<string>("");
   const [phone, setPhone] = useState<string>("");
   const [otpId, setOtpId] = useState<string | null>(null);
@@ -106,10 +105,10 @@ const Auth: React.FC<AuthProps> = ({
   };
 
   useEffect(() => {
-    if (authIframeClient) {
+    if (passkeyIframeClient) {
       setLoading(false);
     }
-  }, [authIframeClient]);
+  }, [passkeyIframeClient]);
 
   if (loading) {
     return (
@@ -136,8 +135,8 @@ const Auth: React.FC<AuthProps> = ({
     expirationSeconds?: string
   ) => {
     if (credentialBundle) {
-      await authIframeClient!.injectCredentialBundle(credentialBundle);
-      await authIframeClient!.loginWithAuthBundle(
+      await passkeyIframeClient!.injectCredentialBundle(credentialBundle);
+      await passkeyIframeClient!.loginWithAuthBundle(
         credentialBundle,
         expirationSeconds
       );
@@ -162,7 +161,7 @@ const Auth: React.FC<AuthProps> = ({
     try {
       if (!passkeyCreated) {
         const { encodedChallenge, attestation } =
-          (await passkeyClient?.createUserPasskey({
+          (await passkeyIframeClient?.createUserPasskey({
             publicKey: { user: { name: siteInfo, displayName: siteInfo } },
           })) || {};
 
@@ -186,15 +185,17 @@ const Auth: React.FC<AuthProps> = ({
         }
       }
 
-      const sessionResponse = await passkeyClient?.createReadWriteSession({
-        targetPublicKey: authIframeClient?.iframePublicKey!,
-        ...(authConfig.sessionLengthSeconds !== undefined && {
-          expirationSeconds: authConfig.sessionLengthSeconds.toString(),
-        }),
-        organizationId:
-          turnkey?.config.defaultOrganizationId ??
-          process.env.NEXT_PUBLIC_ORGANIZATION_ID!,
-      });
+      const sessionResponse = await passkeyIframeClient?.createReadWriteSession(
+        {
+          targetPublicKey: passkeyIframeClient?.iframePublicKey!,
+          ...(authConfig.sessionLengthSeconds !== undefined && {
+            expirationSeconds: authConfig.sessionLengthSeconds.toString(),
+          }),
+          organizationId:
+            turnkey?.config.defaultOrganizationId ??
+            process.env.NEXT_PUBLIC_ORGANIZATION_ID!,
+        }
+      );
       if (sessionResponse?.credentialBundle) {
         await handleAuthSuccess(
           sessionResponse.credentialBundle,
@@ -210,15 +211,17 @@ const Auth: React.FC<AuthProps> = ({
 
   const handleLoginWithPasskey = async () => {
     try {
-      const sessionResponse = await passkeyClient?.createReadWriteSession({
-        targetPublicKey: authIframeClient?.iframePublicKey!,
-        ...(authConfig.sessionLengthSeconds !== undefined && {
-          expirationSeconds: authConfig.sessionLengthSeconds.toString(),
-        }),
-        organizationId:
-          turnkey?.config.defaultOrganizationId ??
-          process.env.NEXT_PUBLIC_ORGANIZATION_ID!,
-      });
+      const sessionResponse = await passkeyIframeClient?.createReadWriteSession(
+        {
+          targetPublicKey: passkeyIframeClient?.iframePublicKey!,
+          ...(authConfig.sessionLengthSeconds !== undefined && {
+            expirationSeconds: authConfig.sessionLengthSeconds.toString(),
+          }),
+          organizationId:
+            turnkey?.config.defaultOrganizationId ??
+            process.env.NEXT_PUBLIC_ORGANIZATION_ID!,
+        }
+      );
 
       if (sessionResponse?.credentialBundle) {
         await handleAuthSuccess(
@@ -262,7 +265,7 @@ const Auth: React.FC<AuthProps> = ({
       otpType,
       contact: value,
       ...(customSmsMessage && { customSmsMessage }),
-      userIdentifier: authIframeClient?.iframePublicKey!,
+      userIdentifier: passkeyIframeClient?.iframePublicKey!,
     });
     if (initAuthResponse && initAuthResponse.otpId) {
       setSuborgId(suborgId!);
@@ -295,7 +298,7 @@ const Auth: React.FC<AuthProps> = ({
     const oauthResponse = await server.oauth({
       suborgID: suborgId!,
       oidcToken: credential,
-      targetPublicKey: authIframeClient?.iframePublicKey!,
+      targetPublicKey: passkeyIframeClient?.iframePublicKey!,
       sessionLengthSeconds: authConfig.sessionLengthSeconds,
     });
     if (oauthResponse && oauthResponse.token) {
@@ -410,7 +413,7 @@ const Auth: React.FC<AuthProps> = ({
               authConfig.googleClientId ??
               process.env.NEXT_PUBLIC_GOOGLE_CLIENT_ID!
             }
-            iframePublicKey={authIframeClient!.iframePublicKey!}
+            iframePublicKey={passkeyIframeClient!.iframePublicKey!}
             onSuccess={(response: any) =>
               handleOAuthLogin(response.idToken, "Google")
             }
@@ -423,7 +426,7 @@ const Auth: React.FC<AuthProps> = ({
               authConfig.appleClientId ??
               process.env.NEXT_PUBLIC_APPLE_CLIENT_ID!
             }
-            iframePublicKey={authIframeClient!.iframePublicKey!}
+            iframePublicKey={passkeyIframeClient!.iframePublicKey!}
             onSuccess={(response: any) =>
               handleOAuthLogin(response.idToken, "Apple")
             }
@@ -436,7 +439,7 @@ const Auth: React.FC<AuthProps> = ({
               authConfig.facebookClientId ??
               process.env.NEXT_PUBLIC_FACEBOOK_CLIENT_ID!
             }
-            iframePublicKey={authIframeClient!.iframePublicKey!}
+            iframePublicKey={passkeyIframeClient!.iframePublicKey!}
             onSuccess={(response: any) =>
               handleOAuthLogin(response.id_token, "Facebook")
             }
@@ -691,7 +694,7 @@ const Auth: React.FC<AuthProps> = ({
                     suborgId={suborgId}
                     otpId={otpId!}
                     sessionLengthSeconds={authConfig.sessionLengthSeconds}
-                    authIframeClient={authIframeClient!}
+                    passkeyIframeClient={passkeyIframeClient!}
                     onValidateSuccess={handleAuthSuccess}
                     onResendCode={handleResendCode}
                   />
